printf

+++ convert_number.c
@@ -14,7 +14,7 @@ int print_hex(va_list ap, params_t *params)
 	if (params->l_modifier)
 		l = (unsigned long)va_arg(ap, unsigned long);
 	else if (params->h_modifier)
-		l = (unsigned short int)va_arg(ap, unsigned short int);
+		l = (unsigned short int)va_arg(ap, int);
 	else
 		l = (unsigned int)va_arg(ap, unsigned int);
 
@@ -42,7 +42,7 @@ int print_HEX(va_list ap, params_t *params)
 	if (params->l_modifier)
 		l = (unsigned long)va_arg(ap, unsigned long);
 	else if (params->h_modifier)
-		l = (unsigned short int)va_arg(ap, unsigned short int);
+		l = (unsigned short int)va_arg(ap, int);
 	else
 		l = (unsigned int)va_arg(ap, unsigned int);
 
@@ -88,7 +88,7 @@ int print_octal(va_list ap, params_t *params)
 	if (params->l_modifier)
 		l = (unsigned long)va_arg(ap, unsigned long);
 	else if (params->h_modifier)
-		l = (unsigned short int)va_arg(ap, unsigned short int);
+		l = (unsigned short int)va_arg(ap, int);
 	else
 		l = (unsigned int)va_arg(ap, unsigned int);
